handle action on qr
